requirements/test.txt
django-debug-toolbar==5.1.0
flake8==7.2.0
flake8-broken-line==1.0.0
flake8-bugbear==24.12.12
flake8-clean-block==0.1.2
flake8-commas==4.0.0
flake8-comments==0.1.2
flake8-comprehensions==3.16.0
flake8-continuation==1.0.5
flake8-dunder-all==0.4.1
flake8-eradicate==1.5.0
flake8-fixme==1.1.1
flake8-forbidden-func==0.1.0
flake8-implicit-str-concat==0.5.0
flake8-import-order==0.18.2
flake8-mutable==1.2.0
flake8-print==5.0.0
flake8-return==1.2.0
flake8-use-pathlib==0.3.0
freezegun==1.5.1
isort==6.0.0
parameterized==0.9.0
pep8-naming==0.14.1
-r prod.txt